fix(mysql): force utf8mb4 on the seed and move parseId out of controllers
The MySQL container ships without a locale, so the `mysql` client that applies
docker/mysql/sql/*.sql negotiates latin1 and read every UTF-8 byte of the seed
as a separate character: "Juárez" was stored as "JuÃ¡rez" in a column that is
utf8mb4. A `SET NAMES utf8mb4` at the top of the seed fixes it at the source.

parse-id.ts was sitting in presentation/controllers/ but it is not a
controller — it is an HTTP-layer helper that turns a route param into a
positive integer or fails with 400. It now lives in presentation/utils/,
alongside its test.
